+++ packages/prisma/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @askdb/prisma
 
+## 0.2.0-beta.9
+
+### Patch Changes
+
+- @askdb/introspect@0.3.0-beta.9
+
+## 0.2.0-beta.8
+
+### Patch Changes
+
+- Updated dependencies [49efa32]
+  - @askdb/introspect@0.3.0-beta.8
+
 ## 0.2.0-beta.7
 
 ### Patch Changes